requirements.txt
# === Core Python Dependencies ===
python-dotenv==1.0.1
requests==2.31.0
tqdm==4.66.2

# === Data Handling & Processing ===
pandas==2.2.2
numpy==1.26.4

# === NLP & Text Embeddings ===
spacy==3.7.4
scispacy==0.5.2
sentence-transformers==2.7.0
transformers==4.40.0
torch==2.3.0

# === Vector Database & Search ===
faiss-cpu==1.8.0

# === Summarization & RAG Support ===
openai==1.23.6
langchain==0.1.16
langchain-community==0.0.33

# === Visualization & Graph Tools ===
networkx==3.3
matplotlib==3.9.0
pyvis==0.3.2
plotly==5.22.0

# === Utility & Logging ===
loguru==0.7.2
rich==13.7.0

# === Export & Reports ===
reportlab==4.2.0

# === Web API ===
fastapi==0.111.0
uvicorn[standard]==0.29.0

# === Testing & Quality ===
pytest==8.2.0
pytest-cov==5.0.0
flake8==7.0.0
black==24.4.2
